Create types file
